pyproject.toml
[tool.poetry]
name = "weekly-report"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
